Move unit tests into own file
